+++ source/Shared/Systems/CameraSystem.cpp
@@ -18,9 +18,9 @@ namespace Game
          glm::tvec2<double> scroll = mInput->GetMouseScroll();
          
          entities.Each<Engine::Transform, MouseControlledCamera>([&](Engine::Entity /*entity*/, Engine::Transform& transform, MouseControlledCamera& opts) {
-            transform.SetYaw(transform.GetYaw() + opts.sensitivity[0] * movement.x);
+            transform.SetYaw(transform.GetYaw() + float(opts.sensitivity[0] * movement.x));
 
-            float newPitch = transform.GetPitch() - opts.sensitivity[1] * movement.y;
+            float newPitch = transform.GetPitch() - float(opts.sensitivity[1] * movement.y);
             if (newPitch < -M_PI / 2.0f + 0.01f)
             {
                newPitch = -M_PI / 2.0f + 0.01f;
@@ -33,7 +33,7 @@ namespace Game
          });
 
          entities.Each<ArmCamera, MouseControlledCameraArm>([&](Engine::Entity /*entity*/, ArmCamera& camera, MouseControlledCameraArm& opts) {
-            camera.distance -= opts.sensitivity * scroll.y;
+            camera.distance -= float(opts.sensitivity * scroll.y);
 
             if (camera.distance < camera.minDistance)
             {
@@ -51,9 +51,9 @@ namespace Game
             int yMove = mInput->IsKeyDown(opts.keys[1]) - mInput->IsKeyDown(opts.keys[0]);
             int xMove = mInput->IsKeyDown(opts.keys[3]) - mInput->IsKeyDown(opts.keys[2]);
 
-            transform.SetYaw(transform.GetYaw() + dt * opts.sensitivity[0] * xMove);
+            transform.SetYaw(transform.GetYaw() + float(dt * opts.sensitivity[0] * xMove));
 
-            float newPitch = transform.GetPitch() - dt * opts.sensitivity[1] * yMove;
+            float newPitch = transform.GetPitch() - float(dt * opts.sensitivity[1] * yMove);
             if (newPitch < -M_PI / 2.0f + 0.01f)
             {
                newPitch = -M_PI / 2.0f + 0.01f;
@@ -67,7 +67,7 @@ namespace Game
 
          entities.Each<ArmCamera, KeyControlledCameraArm>([&](Engine::Entity /*entity*/, ArmCamera& camera, KeyControlledCameraArm& opts) {
             int move = mInput->IsKeyDown(opts.zoomOut) - mInput->IsKeyDown(opts.zoomIn);
-            camera.distance += dt * opts.sensitivity * move;
+            camera.distance += float(dt * opts.sensitivity * move);
 
             if (camera.distance < camera.minDistance)
             {
